Transactions are connected to database

+++ src/modals/Customer.java
@@ -24,7 +24,7 @@ public enum Customer {
         return accountBalance;
     }
 
-    public void setAccountBalance(int accountBalance) {
+    public void setAccountBalance(double accountBalance) {
         this.accountBalance = accountBalance;
     }
